Logique des objets mise en place; Page Équipements terminée à 98%; Se concentrer sur la page principale et tant pis pour la logique de paiement et de poids pour l'instant, ça sera fait manuellement;

+++ maFichePersonnageJDR/Model/NewArmesModel.cs
@@ -178,7 +178,7 @@ namespace maFichePersonnageJDR.Model
                 {
                     while (reader.Read())
                     {
-                        poidsArme = Convert.ToInt32(reader["poids_arme"]).ToString();
+                        poidsArme = Convert.ToDouble(reader["poids_arme"]).ToString();
                     }
                 }
